4-print_rev: printing a string in reverse order

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -6,7 +6,7 @@
  */
 void print_rev(char *s)
 {
-	long n, i = 0;
+	int n, i = 0;
 
 	for (n = 0; s[n] != '\0'; n++)
 	i = n;
